+++ src/WhereGroup/MetadorBundle/Controller/DefaultController.php
@@ -115,6 +115,31 @@ class DefaultController extends Controller
     }
 
     /**
+     * @Route("/obj/{id}")
+     */
+    public function objAction($id) {
+        $em = $this->getDoctrine()->getManager();
+
+        $metadata = $em->getRepository('WhereGroupMetadorBundle:Metadata')->findOneById($id);
+        
+        if($metadata) {
+            $p = unserialize($metadata->getMetadata());
+
+            die('<pre>' . print_r($p, 1) . '</pre>');
+
+        } else {
+            $xml = $this->render("WhereGroupMetadorBundle:Xml:exception.xml.twig", array(
+                "message" => "Datensatz nicht gefunden."
+            ));
+        }
+
+        $response = new Response();
+        $response->headers->set('Content-Type', 'text/xml');
+        $response->setContent($xml->getContent());
+        
+        return $response;
+    }
+    /**
      * @Route("/help/get")
      * @Method({"POST", "GET"})
      */