+++ app/Resources/doc/changelog.md
@@ -3,7 +3,8 @@
 ## 2.0.2.2
 * Moved BBOX config from parameters.yml to MetadorBundle/data folder.
 * Moved metador parameters into metador bundle.
-* Added Alternate Title to dataset and service forms.
+* Added alternate title to dataset and service forms.
+* Added "objAction" to print metadata object.
 
 ## 2.0.2.1
 * Upgrade Symfony version (2.3.7).